refactor: change the geojson parsing of multi-
When parsing a multi polygon, line or point, the result was a modified
Array with custom properties, not respecting the original object. It has
been replaced by an Object containig the Array in question, under the
name elements.

Also, when a multi polygon, line or point, is composed of only one
element, we simplify the result by returning directly a polygon, line or
point.